拍照

+++ example/android/device/src/main/java/com/qc/device/model/Result.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 
 data class Result<T>(
-    @SerializedName("code") val code: Long,
+    @SerializedName("code") val code: Int,
     @SerializedName("message") val message: String? = null,
     @SerializedName("data") val data: T,
 )
